Backend/FastAPI/requirements.txt
# Clase en vídeo: https://youtu.be/_y9qQZXE24A?t=27335
fastapi
python-jose
passlib
bcrypt
pymongo
python-multipart